Refactor: Move `EndpointCategory` into `web_tools.rs`
