+++ crates/re_viewer/src/web.rs
@@ -1,13 +1,13 @@
+//! Main entry-point of the web app.
+
 #![allow(clippy::mem_forget)] // False positives from #[wasm_bindgen] macro
 
-use anyhow::Context as _;
-use eframe::wasm_bindgen::{self, prelude::*};
-
-use std::ops::ControlFlow;
-use std::sync::Arc;
+use wasm_bindgen::prelude::*;
 
 use re_log::ResultExt as _;
 use re_memory::AccountingAllocator;
+
+use crate::web_tools::url_to_receiver;
 
 #[global_allocator]
 static GLOBAL: AccountingAllocator<std::alloc::System> =
@@ -90,7 +90,7 @@ impl WebHandle {
         let Some(mut app) = self.runner.app_mut::<crate::App>() else {
             return;
         };
-        let rx = url_to_receiver(url, app.re_ui.egui_ctx.clone());
+        let rx = url_to_receiver(app.re_ui.egui_ctx.clone(), url);
         if let Some(rx) = rx.ok_or_log_error() {
             app.add_receiver(rx);
         }
@@ -130,8 +130,6 @@ fn create_app(
     };
     let re_ui = crate::customize_eframe(cc);
 
-    let egui_ctx = cc.egui_ctx.clone();
-
     let mut app = crate::App::new(build_info, &app_env, startup_options, re_ui, cc.storage);
 
     let query_map = &cc.integration_info.web_info.location.query_map;
@@ -145,73 +143,19 @@ fn create_app(
     }
 
     if let Some(url) = url {
-        if let Some(receiver) = url_to_receiver(url, egui_ctx).ok_or_log_error() {
+        if let Some(receiver) = url_to_receiver(cc.egui_ctx.clone(), url).ok_or_log_error() {
             app.add_receiver(receiver);
         }
     } else {
         // NOTE: we support passing in multiple urls to multiple different recorording, blueprints, etc
         for url in query_map.get("url").into_iter().flatten() {
-            if let Some(receiver) = url_to_receiver(url, egui_ctx.clone()).ok_or_log_error() {
+            if let Some(receiver) = url_to_receiver(cc.egui_ctx.clone(), url).ok_or_log_error() {
                 app.add_receiver(receiver);
             }
         }
     }
 
     app
-}
-
-fn url_to_receiver(
-    url: &str,
-    egui_ctx: egui::Context,
-) -> anyhow::Result<re_smart_channel::Receiver<re_log_types::LogMsg>> {
-    let ui_waker = Box::new(move || {
-        // Spend a few more milliseconds decoding incoming messages,
-        // then trigger a repaint (https://github.com/rerun-io/rerun/issues/963):
-        egui_ctx.request_repaint_after(std::time::Duration::from_millis(10));
-    });
-    match categorize_uri(url) {
-        EndpointCategory::HttpRrd(url) => Ok(
-            re_log_encoding::stream_rrd_from_http::stream_rrd_from_http_to_channel(
-                url,
-                Some(ui_waker),
-            ),
-        ),
-        EndpointCategory::WebEventListener => {
-            // Process an rrd when it's posted via `window.postMessage`
-            let (tx, rx) = re_smart_channel::smart_channel(
-                re_smart_channel::SmartMessageSource::RrdWebEventCallback,
-                re_smart_channel::SmartChannelSource::RrdWebEventListener,
-            );
-            re_log_encoding::stream_rrd_from_http::stream_rrd_from_event_listener(Arc::new({
-                move |msg| {
-                    ui_waker();
-                    use re_log_encoding::stream_rrd_from_http::HttpMessage;
-                    match msg {
-                        HttpMessage::LogMsg(msg) => {
-                            if tx.send(msg).is_ok() {
-                                ControlFlow::Continue(())
-                            } else {
-                                re_log::info!("Failed to send log message to viewer - closing");
-                                ControlFlow::Break(())
-                            }
-                        }
-                        HttpMessage::Success => {
-                            tx.quit(None).warn_on_err_once("failed to send quit marker");
-                            ControlFlow::Break(())
-                        }
-                        HttpMessage::Failure(err) => {
-                            tx.quit(Some(err))
-                                .warn_on_err_once("failed to send quit marker");
-                            ControlFlow::Break(())
-                        }
-                    }
-                }
-            }));
-            Ok(rx)
-        }
-        EndpointCategory::WebSocket(url) => re_data_source::connect_to_ws_url(&url, Some(ui_waker))
-            .with_context(|| format!("Failed to connect to WebSocket server at {url}.")),
-    }
 }
 
 /// Used to set the "email" property in the analytics config,
@@ -225,38 +169,6 @@ pub fn set_email(email: String) {
     let mut config = re_analytics::Config::load().unwrap().unwrap_or_default();
     config.opt_in_metadata.insert("email".into(), email.into());
     config.save().unwrap();
-}
-
-enum EndpointCategory {
-    /// Could be a local path (`/foo.rrd`) or a remote url (`http://foo.com/bar.rrd`).
-    ///
-    /// Could be a link to either an `.rrd` recording or a `.rbl` blueprint.
-    HttpRrd(String),
-
-    /// A remote Rerun server.
-    WebSocket(String),
-
-    /// An eventListener for rrd posted from containing html
-    WebEventListener,
-}
-
-fn categorize_uri(uri: &str) -> EndpointCategory {
-    if uri.starts_with("http") || uri.ends_with(".rrd") || uri.ends_with(".rbl") {
-        EndpointCategory::HttpRrd(uri.into())
-    } else if uri.starts_with("ws:") || uri.starts_with("wss:") {
-        EndpointCategory::WebSocket(uri.into())
-    } else if uri.starts_with("web_event:") {
-        EndpointCategory::WebEventListener
-    } else {
-        // If this is something like `foo.com` we can't know what it is until we connect to it.
-        // We could/should connect and see what it is, but for now we just take a wild guess instead:
-        re_log::info!("Assuming WebSocket endpoint");
-        if uri.contains("://") {
-            EndpointCategory::WebSocket(uri.into())
-        } else {
-            EndpointCategory::WebSocket(format!("{}://{uri}", re_ws_comms::PROTOCOL))
-        }
-    }
 }
 
 fn is_in_notebook(info: &eframe::IntegrationInfo) -> bool {
