+++ crates/re_viewer/src/web_tools.rs
@@ -1,3 +1,11 @@
+use std::{ops::ControlFlow, sync::Arc};
+
+use anyhow::Context as _;
+
+use re_log::ResultExt as _;
+
+/// Web-specific tools used by various parts of the application.
+
 pub fn set_url_parameter_and_refresh(key: &str, value: &str) -> Result<(), wasm_bindgen::JsValue> {
     let Some(window) = web_sys::window() else {
         return Err("Failed to get window".into());
@@ -8,4 +16,94 @@ pub fn set_url_parameter_and_refresh(key: &str, value: &str) -> Result<(), wasm_
     url.search_params().set(key, value);
 
     location.assign(&url.href())
+}
+
+enum EndpointCategory {
+    /// Could be a local path (`/foo.rrd`) or a remote url (`http://foo.com/bar.rrd`).
+    ///
+    /// Could be a link to either an `.rrd` recording or a `.rbl` blueprint.
+    HttpRrd(String),
+
+    /// A remote Rerun server.
+    WebSocket(String),
+
+    /// An eventListener for rrd posted from containing html
+    WebEventListener,
+}
+
+impl EndpointCategory {
+    fn categorize_uri(uri: &str) -> Self {
+        if uri.starts_with("http") || uri.ends_with(".rrd") || uri.ends_with(".rbl") {
+            Self::HttpRrd(uri.into())
+        } else if uri.starts_with("ws:") || uri.starts_with("wss:") {
+            Self::WebSocket(uri.into())
+        } else if uri.starts_with("web_event:") {
+            Self::WebEventListener
+        } else {
+            // If this is something like `foo.com` we can't know what it is until we connect to it.
+            // We could/should connect and see what it is, but for now we just take a wild guess instead:
+            re_log::info!("Assuming WebSocket endpoint");
+            if uri.contains("://") {
+                Self::WebSocket(uri.into())
+            } else {
+                Self::WebSocket(format!("{}://{uri}", re_ws_comms::PROTOCOL))
+            }
+        }
+    }
+}
+
+/// Start receiving from the given url.
+pub fn url_to_receiver(
+    egui_ctx: egui::Context,
+    url: &str,
+) -> anyhow::Result<re_smart_channel::Receiver<re_log_types::LogMsg>> {
+    let ui_waker = Box::new(move || {
+        // Spend a few more milliseconds decoding incoming messages,
+        // then trigger a repaint (https://github.com/rerun-io/rerun/issues/963):
+        egui_ctx.request_repaint_after(std::time::Duration::from_millis(10));
+    });
+    match EndpointCategory::categorize_uri(url) {
+        EndpointCategory::HttpRrd(url) => Ok(
+            re_log_encoding::stream_rrd_from_http::stream_rrd_from_http_to_channel(
+                url,
+                Some(ui_waker),
+            ),
+        ),
+        EndpointCategory::WebEventListener => {
+            // Process an rrd when it's posted via `window.postMessage`
+            let (tx, rx) = re_smart_channel::smart_channel(
+                re_smart_channel::SmartMessageSource::RrdWebEventCallback,
+                re_smart_channel::SmartChannelSource::RrdWebEventListener,
+            );
+            let url = url.to_owned();
+            re_log_encoding::stream_rrd_from_http::stream_rrd_from_event_listener(Arc::new({
+                move |msg| {
+                    ui_waker();
+                    use re_log_encoding::stream_rrd_from_http::HttpMessage;
+                    match msg {
+                        HttpMessage::LogMsg(msg) => {
+                            if tx.send(msg).is_ok() {
+                                ControlFlow::Continue(())
+                            } else {
+                                re_log::info_once!("Failed to send log message to viewer - closing connection to {url}");
+                                ControlFlow::Break(())
+                            }
+                        }
+                        HttpMessage::Success => {
+                            tx.quit(None).warn_on_err_once("Failed to send quit marker");
+                            ControlFlow::Break(())
+                        }
+                        HttpMessage::Failure(err) => {
+                            tx.quit(Some(err))
+                                .warn_on_err_once("Failed to send quit marker");
+                            ControlFlow::Break(())
+                        }
+                    }
+                }
+            }));
+            Ok(rx)
+        }
+        EndpointCategory::WebSocket(url) => re_data_source::connect_to_ws_url(&url, Some(ui_waker))
+            .with_context(|| format!("Failed to connect to WebSocket server at {url}.")),
+    }
 }
